requirements.txt
python-telegram-bot
yt_dlp
mutagen
django
requests
ConfigArgParse
python-dotenv
